% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_miniconda_is_installed (copy).R
\name{check_miniconda_is_installed}
\alias{check_miniconda_is_installed}
\title{Check if Miniconda is installed}
\usage{
check_miniconda_is_installed(ormr_folder_name)
}
\arguments{
\item{ormr_folder_name}{name of the folder where \code{ormr}
installs all Python environments and packages}
}
\value{
Nothing
}
\description{
Check if Miniconda is installed.
Will \link{stop} if not
}
\author{
Richèl J.C. Bilderbeek
}
