% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pip_is_installed.R
\name{check_pip_is_installed}
\alias{check_pip_is_installed}
\title{Check if \code{pip} is installed}
\usage{
check_pip_is_installed(ormr_folder_name)
}
\arguments{
\item{ormr_folder_name}{name of the folder where \code{ormr}
installs all Python environments and packages}
}
\value{
Nothing
}
\description{
Check if \code{pip} is installed.
Will \link{stop} if not
}
\examples{
if (is_pip_installed()) {
  check_pip_is_installed
}
}
\author{
Richèl J.C. Bilderbeek
}
