% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_python_package_name.R
\name{check_python_package_name}
\alias{check_python_package_name}
\title{Check if the Python package name is valid}
\usage{
check_python_package_name(package_name)
}
\arguments{
\item{package_name}{the name of one Python package}
}
\value{
Nothing
}
\description{
Check if the Python package name is valid
Will \link{stop} if not
}
\examples{
if (!plinkr::is_on_ci()) {
  check_python_package_name("scipy")
}
}
\author{
Richèl J.C. Bilderbeek
}
