% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_and_activate_conda_env.R
\name{create_and_activate_conda_env}
\alias{create_and_activate_conda_env}
\title{Creates and activates the Conda environment}
\usage{
create_and_activate_conda_env(
  ormr_folder_name,
  python_version = get_default_python_version(),
  verbose = FALSE
)
}
\arguments{
\item{ormr_folder_name}{name of the folder where \code{ormr}
installs all Python environments and packages}

\item{python_version}{the version of Python used.
Use \link{get_default_python_version} to get
the default Python version.}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link[plinkr]{check_verbose} to detect if this argument is valid.}
}
\value{
Nothing
}
\description{
Creates and activates the Conda environment.
If the Conda environment already exists, nothing new will happen
}
\author{
Richèl J.C. Bilderbeek
}
