% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_conda_env.R
\name{create_conda_env}
\alias{create_conda_env}
\title{Create a Conda environment}
\usage{
create_conda_env(ormr_folder_name, python_version = "3.6", verbose = FALSE)
}
\arguments{
\item{ormr_folder_name}{name of the folder where \code{ormr}
installs all Python environments and packages}

\item{python_version}{the Python version,
which is 3.6 by default
(thanks to \url{https://stackoverflow.com/a/69978354}),
as \code{ormr} is a sub-package
for \code{gcaer}, where the latter needs to install the Python
\code{tensorflow} package. For more recent Python versions, this
install fails.}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link[plinkr]{check_verbose} to detect if this argument is valid.}
}
\value{
Nothing
}
\description{
Create a Conda environment
}
\note{
the output of this function cannot be silenced, thanks
to \link[reticulate]{install_miniconda}
}
\examples{
# Install to a temporary folder
ormr_folder_name <- tempfile()
create_conda_env(ormr_folder_name = ormr_folder_name)

# Clean up
unlink(ormr_folder_name, recursive = TRUE)
}
\author{
Richèl J.C. Bilderbeek
}
