% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_default_conda_env.R
\name{create_default_conda_env}
\alias{create_default_conda_env}
\title{Create and/or get the default \code{ormr} Conda enviroment}
\usage{
create_default_conda_env(verbose = FALSE)
}
\value{
the default \code{ormr} folder name
}
\description{
Create and/or get the default \code{ormr} Conda enviroment
}
\examples{
create_default_conda_env()
}
\author{
Richèl J.C. Bilderbeek
}
