% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_installed_python_packages.R
\name{get_installed_python_packages}
\alias{get_installed_python_packages}
\title{Get the installed  Python packages and their versions}
\usage{
get_installed_python_packages(ormr_folder_name, verbose = FALSE)
}
\arguments{
\item{ormr_folder_name}{name of the folder where \code{ormr}
installs all Python environments and packages}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link[plinkr]{check_verbose} to detect if this argument is valid.}
}
\value{
a \link[tibble]{tibble} with two rows:
\itemize{
\item \code{package}: the package name
\item \code{version}: the package version
\item \code{requirement}: the packages requirements
\item \code{channel}: the package's channel
}
}
\description{
Get the installed  Python packages and their versions
}
\author{
Richèl J.C. Bilderbeek
}
