% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_miniconda_path.R
\name{get_miniconda_path}
\alias{get_miniconda_path}
\title{Get the path where Miniconda and all Pyton packages will be installed}
\usage{
get_miniconda_path(ormr_folder_name)
}
\arguments{
\item{ormr_folder_name}{name of the folder where \code{ormr}
installs all Python environments and packages}
}
\value{
the \code{miniconda_path}, as used by \link[reticulate]{reticulate}
}
\description{
Get the path where Miniconda and all Pyton packages will be installed
}
\examples{
get_miniconda_path(ormr_folder_name = tempfile())
}
\author{
Richèl J.C. Bilderbeek
}
