% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_python_binary_path.R
\name{get_python_binary_path}
\alias{get_python_binary_path}
\title{Get the path to the \code{python} binary}
\usage{
get_python_binary_path(ormr_folder_name)
}
\arguments{
\item{ormr_folder_name}{name of the folder where \code{ormr}
installs all Python environments and packages}
}
\value{
the path to the actve \code{python} binary
}
\description{
Get the path to the \code{python} binary
}
\author{
Richèl J.C. Bilderbeek
}
