% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_miniconda.R
\name{install_miniconda}
\alias{install_miniconda}
\title{Install Miniconda}
\usage{
install_miniconda(ormr_folder_name)
}
\arguments{
\item{ormr_folder_name}{name of the folder where \code{ormr}
installs all Python environments and packages}
}
\value{
Nothing
}
\description{
Install Miniconda
}
\note{
the output of this function cannot be silenced, thanks
to \link[reticulate]{install_miniconda}
}
\examples{
# Install to a temporary folder
ormr_folder_name <- tempfile()
install_miniconda(ormr_folder_name = ormr_folder_name)

# Clean up
unlink(ormr_folder_name, recursive = TRUE)
}
\author{
Richèl J.C. Bilderbeek
}
