% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_pip.R
\name{install_pip}
\alias{install_pip}
\title{Install \code{pip}}
\usage{
install_pip(ormr_folder_name = create_ormr_folder_name(), verbose = FALSE)
}
\arguments{
\item{ormr_folder_name}{name of the folder where \code{ormr}
installs all Python environments and packages}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link[plinkr]{check_verbose} to detect if this argument is valid.}
}
\value{
Nothing
}
\description{
Install \code{pip}
}
\examples{
is_pip_installed()
}
\author{
Richèl J.C. Bilderbeek
}
