% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_python_package.R
\name{install_python_package}
\alias{install_python_package}
\title{Install one Python package}
\usage{
install_python_package(ormr_folder_name, package_name, verbose = FALSE)
}
\arguments{
\item{ormr_folder_name}{name of the folder where \code{ormr}
installs all Python environments and packages}

\item{package_name}{the name of one Python package}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link[plinkr]{check_verbose} to detect if this argument is valid.}
}
\value{
nothing
}
\description{
Install one Python package
}
\examples{
ormr_folder_name <- create_default_conda_env()

# Note that this package may already be installed,
# but it works the same for any other package
install_python_package(
  ormr_folder_name = ormr_folder_name,
  package_name = "scipy"
)
}
\seealso{
use \link{install_python_packages} to install one
or more Python packages
}
\author{
Richèl J.C. Bilderbeek
}
