% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_conda_installed.R
\name{is_conda_installed}
\alias{is_conda_installed}
\title{Determine if Conda is installed}
\usage{
is_conda_installed(folder_name)
}
\arguments{
\item{folder_name}{name of the folder where all Python
environments and packages are installed}
}
\description{
Determine if Conda is installed
}
