% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_miniconda_installed.R
\name{is_miniconda_installed}
\alias{is_miniconda_installed}
\title{Determine if Conda is installed in the \code{ormr_folder_name}}
\usage{
is_miniconda_installed(ormr_folder_name)
}
\arguments{
\item{ormr_folder_name}{name of the folder where \code{ormr}
installs all Python environments and packages}
}
\description{
Determine if Conda is installed in the \code{ormr_folder_name}
}
