% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_pip_installed.R
\name{is_pip_installed}
\alias{is_pip_installed}
\title{Determine if \code{pip} is installed}
\usage{
is_pip_installed(ormr_folder_name, verbose = FALSE)
}
\arguments{
\item{ormr_folder_name}{name of the folder where \code{ormr}
installs all Python environments and packages}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link[plinkr]{check_verbose} to detect if this argument is valid.}
}
\value{
TRUE if \code{pip} is installed
}
\description{
Determine if \code{pip} is installed
}
\examples{
is_pip_installed(ormr_folder_name = tempfile())
}
\author{
Richèl J.C. Bilderbeek
}
