% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_python_package_installed.R
\name{is_python_package_installed}
\alias{is_python_package_installed}
\title{Determine if a Python package is installed}
\usage{
is_python_package_installed(
  package_name,
  ormr_folder_name = get_default_ormr_folder_name(),
  python_version = get_default_python_version(),
  verbose = FALSE
)
}
\arguments{
\item{package_name}{the Python package's name}

\item{ormr_folder_name}{can be either \code{python3}
or the name of the folder where \code{ormr}
installs all Python environments and packages.}

\item{python_version}{the version of Python used.
Use \link{get_default_python_version} to get
the default Python version.}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link[plinkr]{check_verbose} to detect if this argument is valid.}
}
\value{
nothing
}
\description{
Determine if a Python package is installed
}
\examples{
if (!plinkr::is_on_ci()) {
  ormr_folder_name <- create_default_conda_env()

  # setuptools is installed with the Conda environment
  is_python_package_installed(
    ormr_folder_name = ormr_folder_name,
    package_name = "setuptools"
  )

  # scipy is not installed with the Conda environment
  # Use 'ormr::install_python_packages' to install scipy
  is_python_package_installed(
    ormr_folder_name = ormr_folder_name,
    package_name = "scipy"
  )

}
}
\author{
Richèl J.C. Bilderbeek
}
