% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_python_packages.R
\name{list_python_packages}
\alias{list_python_packages}
\title{List the Python packages and their versions}
\usage{
list_python_packages(ormr_folder_name, verbose = FALSE)
}
\arguments{
\item{ormr_folder_name}{name of the folder where \code{ormr}
installs all Python environments and packages}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link[plinkr]{check_verbose} to detect if this argument is valid.}
}
\value{
a \link[tibble]{tibble} with two rows:
\itemize{
\item \code{package}: the package name
\item \code{version}: the package version
}
}
\description{
List the Python packages and their versions
}
\author{
Richèl J.C. Bilderbeek
}
