% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_python_packages_from_python_bin.R
\name{list_python_packages_from_python_bin}
\alias{list_python_packages_from_python_bin}
\title{List the Python packages and their versions}
\usage{
list_python_packages_from_python_bin(python_bin, verbose = FALSE)
}
\arguments{
\item{python_bin}{path to the Python binary}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link[plinkr]{check_verbose} to detect if this argument is valid.}
}
\value{
a \link[tibble]{tibble} with two rows:
\itemize{
\item \code{package}: the package name
\item \code{version}: the package version
}
}
\description{
List the Python packages and their versions
}
\seealso{
consider using the superior
\code{reticulate:::conda_list_packages(envname = ormr_folder_name)}
}
\author{
Richèl J.C. Bilderbeek
}
