% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ormr_report.R
\name{ormr_report}
\alias{ormr_report}
\title{Create a \code{ormr} report, to be used when reporting bugs}
\usage{
ormr_report(ormr_folder_name)
}
\arguments{
\item{ormr_folder_name}{name of the folder where \code{ormr}
installs all Python environments and packages}
}
\description{
Create a \code{ormr} report, to be used when reporting bugs
}
\examples{
ormr_report(ormr_folder_name = tempfile())
}
\author{
Richèl J.C. Bilderbeek
}
