% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_python_script.R
\name{run_python_script}
\alias{run_python_script}
\title{Run a Python script}
\usage{
run_python_script(
  ormr_folder_name,
  python_script_path,
  python_version = get_default_python_version()
)
}
\arguments{
\item{ormr_folder_name}{name of the folder where \code{ormr}
installs all Python environments and packages}

\item{python_script_path}{path to a Python script}
}
\value{
The output of the Python script
}
\description{
Run a Python script
}
\examples{
ormr_folder_name <- create_default_conda_env()
python_script_path <- system.file(
  "extdata", "hello_world.py", package = "ormr"
)
run_python_script(
  ormr_folder_name = ormr_folder_name,
  python_script_path = python_script_path
)
}
\seealso{
use \link{run_python_script_with_args} to run a Python
script with command-line arguments
}
\author{
Richèl J.C. Bilderbeek
}
