% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_python_script_with_args.R
\name{run_python_script_with_args}
\alias{run_python_script_with_args}
\title{Run a Python script with arguments}
\usage{
run_python_script_with_args(
  ormr_folder_name,
  python_script_path,
  args = c(),
  verbose = FALSE
)
}
\arguments{
\item{ormr_folder_name}{name of the folder where \code{ormr}
installs all Python environments and packages}

\item{python_script_path}{path to a Python script}

\item{args}{arguments to a script, for example \code{--help}}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link[plinkr]{check_verbose} to detect if this argument is valid.}
}
\value{
The output of the Python script
}
\description{
Run a Python script with arguments
}
\author{
Richèl J.C. Bilderbeek
}
