% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version_to_semversion.R
\name{version_to_semversion}
\alias{version_to_semversion}
\title{Convert a version number to a \link[semver]{semver} version number}
\usage{
version_to_semversion(package_version)
}
\arguments{
\item{package_version}{a Python package's version, e.g.
\verb{==3.1}, or \verb{<=2.7.1} or \verb{>=42}}
}
\value{
a version number that is compatible with \link[semver]{semver}
}
\description{
Convert a version number to a \link[semver]{semver} version number
}
\examples{
version_to_semversion("1.2.3") # "1.2.3"
version_to_semversion("1.2") # "1.2.0"
version_to_semversion("1") # "1.0.0"
}
\author{
Richèl J.C. Bilderbeek
}
