% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_conda_env_exists.R
\name{check_conda_env_exists}
\alias{check_conda_env_exists}
\title{Check if the Conda environment exists}
\usage{
check_conda_env_exists(ormr_folder_name = get_default_ormr_folder_name())
}
\arguments{
\item{ormr_folder_name}{name of the folder where \code{ormr}
installs all Python environments and packages}
}
\value{
Nothing
}
\description{
Check if the Conda environment exists
Will \link{stop} if not
}
\examples{
check_conda_env_exists()
}
\author{
Richèl J.C. Bilderbeek
}
