% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_default_conda_env.R
\name{create_default_conda_env}
\alias{create_default_conda_env}
\title{Create and/or get the default \code{ormr} Conda enviroment}
\usage{
create_default_conda_env(
  ormr_folder_name = get_default_ormr_folder_name(),
  python_version = get_default_python_version(),
  verbose = FALSE
)
}
\arguments{
\item{ormr_folder_name}{can be either \code{python3}
or the name of the folder where \code{ormr}
installs all Python environments and packages.}

\item{python_version}{the version of Python used.
Use \link{get_default_python_version} to get
the default Python version.}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link[plinkr]{check_verbose} to detect if this argument is valid.}
}
\value{
the default \code{ormr} folder name
}
\description{
Create and/or get the default \code{ormr} Conda enviroment
}
\examples{
create_default_conda_env()
}
\author{
Richèl J.C. Bilderbeek
}
