% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_python_package.R
\name{install_python_package}
\alias{install_python_package}
\title{Install one Python package}
\usage{
install_python_package(
  package_name,
  channel = "",
  ormr_folder_name = get_default_ormr_folder_name(),
  python_version = get_default_python_version(),
  verbose = FALSE
)
}
\arguments{
\item{package_name}{the name of one Python package}

\item{channel}{the channel to install Python packages from,
e.g. \code{bioconda}}

\item{ormr_folder_name}{name of the folder where \code{ormr}
installs all Python environments and packages}

\item{python_version}{the version of Python used.
Use \link{get_default_python_version} to get
the default Python version.}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link[plinkr]{check_verbose} to detect if this argument is valid.}
}
\value{
nothing
}
\description{
Install one Python package
}
\examples{
install_python_package(
  package_name = "scipy"
)
}
\seealso{
use \link{install_python_packages} to install one
or more Python packages
}
\author{
Richèl J.C. Bilderbeek
}
