% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_python_packages.R
\name{install_python_packages}
\alias{install_python_packages}
\title{Install one or more Python packages}
\usage{
install_python_packages(
  ormr_folder_name,
  package_names,
  python_version = "3.6",
  verbose = FALSE
)
}
\arguments{
\item{ormr_folder_name}{name of the folder where \code{ormr}
installs all Python environments and packages}

\item{package_names}{the names of one or more Python packages}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link[plinkr]{check_verbose} to detect if this argument is valid.}
}
\value{
nothing
}
\description{
Install one or more Python packages
}
\examples{
ormr_folder_name <- create_default_conda_env()

# Note that these packages may already be installed
install_python_packages(
  ormr_folder_name = ormr_folder_name,
  package_names = c("scipy", "readline")
)
}
\seealso{
use \link{install_python_package} to install one
Python package
}
\author{
Richèl J.C. Bilderbeek
}
