% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_python_packages.R
\name{install_python_packages}
\alias{install_python_packages}
\title{Install one or more Python packages}
\usage{
install_python_packages(
  package_names,
  ormr_folder_name = get_default_ormr_folder_name(),
  python_version = get_default_python_version(),
  verbose = FALSE
)
}
\arguments{
\item{package_names}{the names of one or more Python packages}

\item{ormr_folder_name}{name of the folder where \code{ormr}
installs all Python environments and packages}

\item{python_version}{the version of Python used.
Use \link{get_default_python_version} to get
the default Python version.}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link[plinkr]{check_verbose} to detect if this argument is valid.}
}
\value{
nothing
}
\description{
Install one or more Python packages
}
\examples{
install_python_packages(
  package_names = c("scipy", "readline")
)
}
\seealso{
use \link{install_python_package} to install one
Python package
}
\author{
Richèl J.C. Bilderbeek
}
