% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_python_packages.R
\name{install_python_packages}
\alias{install_python_packages}
\title{Install one or more Python packages}
\usage{
install_python_packages(
  package_names,
  channel = character(),
  ormr_folder_name = get_default_ormr_folder_name(),
  python_version = get_default_python_version(),
  verbose = FALSE
)
}
\arguments{
\item{package_names}{the names of one or more Python packages}

\item{channel}{the channel to install Python packages from,
e.g. \code{bioconda}}

\item{ormr_folder_name}{can be either \code{python3}
or the name of the folder where \code{ormr}
installs all Python environments and packages.}

\item{python_version}{the version of Python used.
Use \link{get_default_python_version} to get
the default Python version.}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link[plinkr]{check_verbose} to detect if this argument is valid.}
}
\value{
nothing
}
\description{
Install one or more Python packages
}
\examples{
if (plinkr::is_on_ci()) {

  # Use local python
 install_python_packages(
   package_names = c("scipy", "readline"),
   ormr_folder_name = "python3"
 )

  # Use Conda
  if (is_conda_installed()) {
    install_python_packages(
      package_names = c("scipy", "readline")
    )
  }
}
}
\seealso{
use \link{install_python_package} to install one
Python package
}
\author{
Richèl J.C. Bilderbeek
}
