% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_conda_installed.R
\name{is_conda_installed}
\alias{is_conda_installed}
\title{Determine if Conda is installed}
\usage{
is_conda_installed()
}
\value{
TRUE if Conda is installed, FALSE otherwise
}
\description{
Determine if Conda is installed
}
\examples{
is_conda_installed()
}
\author{
Richèl J.C. Bilderbeek
}
