% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_python_package_installed.R
\name{is_python_package_installed}
\alias{is_python_package_installed}
\title{Determine if a Python package is installed}
\usage{
is_python_package_installed(ormr_folder_name, package_name)
}
\arguments{
\item{ormr_folder_name}{name of the folder where \code{ormr}
installs all Python environments and packages}
}
\value{
nothing
}
\description{
Determine if a Python package is installed
}
\author{
Richèl J.C. Bilderbeek
}
