% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_python_package_with_version_installed.R
\name{is_python_package_with_version_installed}
\alias{is_python_package_with_version_installed}
\title{Determine if a Python package with the right version is installed}
\usage{
is_python_package_with_version_installed(
  ormr_folder_name,
  package_name,
  package_version
)
}
\arguments{
\item{ormr_folder_name}{name of the folder where \code{ormr}
installs all Python environments and packages}

\item{package_name}{a Python package's version, e.g.
\verb{==3.1}, or \verb{<=2.7.1} or \verb{>=42}}
}
\value{
nothing
}
\description{
Determine if a Python package with the right version is installed
}
\examples{
if (!plinkr::is_on_ci()) {
  ormr_folder_name <- create_default_conda_env()

  # setuptools is installed with the Conda environment
  is_python_package_installed(
    ormr_folder_name = ormr_folder_name,
    package_name = "setuptools"
  )

  # scipy is not installed with the Conda environment
  # Use 'ormr::install_python_packages' to install scipy
  is_python_package_installed(
    ormr_folder_name = ormr_folder_name,
    package_name = "scipy"
  )

}
}
\author{
Richèl J.C. Bilderbeek
}
