% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ormr_report.R
\name{ormr_report}
\alias{ormr_report}
\title{Create a \code{ormr} report, to be used when reporting bugs}
\usage{
ormr_report(
  ormr_folder_name = get_default_ormr_folder_name(),
  python_version = get_default_python_version(),
  verbose = FALSE
)
}
\arguments{
\item{ormr_folder_name}{can be either \code{python3}
or the name of the folder where \code{ormr}
installs all Python environments and packages.}

\item{python_version}{the version of Python used.
Use \link{get_default_python_version} to get
the default Python version.}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link[plinkr]{check_verbose} to detect if this argument is valid.}
}
\description{
Create a \code{ormr} report, to be used when reporting bugs
}
\examples{
if (!plinkr::is_on_ci()) {
  ormr_report()
}
}
\author{
Richèl J.C. Bilderbeek
}
