% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_python_script.R
\name{run_python_script}
\alias{run_python_script}
\title{Run a Python script}
\usage{
run_python_script(
  python_script_path,
  ormr_folder_name = get_default_ormr_folder_name(),
  python_version = get_default_python_version(),
  verbose = FALSE
)
}
\arguments{
\item{python_script_path}{path to a Python script}

\item{ormr_folder_name}{can be either \code{python3}
or the name of the folder where \code{ormr}
installs all Python environments and packages.}

\item{python_version}{the version of Python used.
Use \link{get_default_python_version} to get
the default Python version.}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link[plinkr]{check_verbose} to detect if this argument is valid.}
}
\value{
The output of the Python script
}
\description{
Run a Python script
}
\examples{
if (plinkr::is_on_ci()) {
  if (is_conda_installed()) {
    # Use conda
    ormr_folder_name <- create_default_conda_env()
  } else {
    # Use local python
    ormr_folder_name <- "python3"
  }

  python_script_path <- system.file(
    "extdata", "hello_world.py", package = "ormr"
  )
  run_python_script(
    ormr_folder_name = ormr_folder_name,
    python_script_path = python_script_path
  )
}
}
\seealso{
use \link{run_python_script_with_args} to run a Python
script with command-line arguments
}
\author{
Richèl J.C. Bilderbeek
}
