% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_python_script.R
\name{run_python_script}
\alias{run_python_script}
\title{Run a Python script}
\usage{
run_python_script(ormr_folder_name, python_script_path)
}
\arguments{
\item{ormr_folder_name}{name of the folder where \code{ormr}
installs all Python environments and packages}

\item{python_script_path}{path to a Python script}
}
\value{
Nothing
}
\description{
Run a Python script
}
\author{
Richèl J.C. Bilderbeek
}
