% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_python_script_with_args.R
\name{run_python_script_with_args}
\alias{run_python_script_with_args}
\title{Run a Python script with command-line arguments}
\usage{
run_python_script_with_args(
  python_script_path,
  args = c(),
  ormr_folder_name = get_default_ormr_folder_name(),
  python_version = get_default_python_version(),
  verbose = FALSE
)
}
\arguments{
\item{python_script_path}{path to a Python script}

\item{args}{arguments to a script, for example \code{--help}}

\item{ormr_folder_name}{name of the folder where \code{ormr}
installs all Python environments and packages}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link[plinkr]{check_verbose} to detect if this argument is valid.}
}
\value{
The output of the Python script
}
\description{
Run a Python script with command-line arguments
}
\examples{
ormr_folder_name <- create_default_conda_env()

python_script_path <- system.file(
  "extdata", "show_args.py", package = "ormr"
)
run_python_script_with_args(
  ormr_folder_name = ormr_folder_name,
  python_script_path = python_script_path,
  args = c("Hello", "world")
)
}
\seealso{
use \link{run_python_script} to run a Python
script without command-line arguments
}
\author{
Richèl J.C. Bilderbeek
}
